+++ CHANGELOG.md
@@ -1,3 +1,7 @@
+# Unreleased
+
+- Add AccordionBasic component
+
 # v2.0.0-dev.14
 
 - (InputField, Input, TextareaField, Textarea): Fix innerRef